create AccordionControl button that can be used independently
